[fix#Subscription] 회원의 구독한 강사 리스트 조회 API 수정

See merge request s13-webmobile1-sub1/S13P11E104!151

+++ Backend/src/main/java/com/e104/reciplay/user/subscription/service/SubscriptionQueryService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface SubscriptionQueryService {
     List<Subscription> querySubscriptionsByUserId(Long userId);
-    Long countSubscribers(Long instructorId);
+    Integer countSubscribers(Long instructorId);
     boolean isSubscribedInstructor(Long instructorId, Long userId);
 
 }
